# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Evidence-Grounded Research Workbench API
  version: 0.2.0
  description: Full MVP API contract for the evidence-grounded multi-model long-horizon research workbench.
servers:
  - url: http://localhost:4000
    description: Local Fastify API
...
paths:
...
  /projects/{projectId}/idea-versions:
    get:
      summary: List idea versions
      tags: [IdeaVersions]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Idea versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaVersionListResponse'
    post:
      summary: Create manual idea version
      tags: [IdeaVersions]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdeaVersionRequest'
      responses:
        '201':
          description: Created idea version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaVersionResponse'
  /idea-versions/{ideaVersionId}:
    get:
      summary: Get idea version with claims and linked critiques
      tags: [IdeaVersions]
      parameters:
        - $ref: '#/components/parameters/IdeaVersionId'
      responses:
        '200':
          description: Idea version detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaVersionDetailResponse'
  /idea-versions/{ideaVersionId}/extract-claims:
    post:
      summary: Extract claims for idea version
      tags: [Claims]
      parameters:
        - $ref: '#/components/parameters/IdeaVersionId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractClaimsRequest'
      responses:
        '201':
          description: Claims extracted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimListResponse'
...
  /projects/{projectId}/presence:
    get:
      tags: [Realtime]
      summary: Get active users for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Presence list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceListResponse'
    post:
      tags: [Realtime]
      summary: Update user presence heartbeat
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userName, page]
              properties:
                userName: { type: string }
                page: { type: string }
      responses:
        '200':
          description: Presence updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      ok:
                        type: boolean
...
  /projects/{projectId}/broadcast:
    post:
      tags: [Realtime]
      summary: Broadcast event to all connected SSE clients
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event, data]
              properties:
                event: { type: string }
                data: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Event broadcast
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      broadcast:
                        type: boolean

  # ── Run Lifecycle (Pause/Resume) ──────────────────────────────────────────
...
  /idea-versions/compare:
    get:
      tags: [IdeaVersions]
      summary: Compare two idea versions
      parameters:
        - name: v1
          in: query
          required: true
          schema: { type: string }
        - name: v2
          in: query
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Version comparison
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
        '400':
          description: Missing v1 or v2 parameters
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    IdeaVersionId:
      name: ideaVersionId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    IdeaVersion:
      type: object
      required: [id, projectId, versionNumber, title, description, status, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        versionNumber: { type: integer }
        title: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [draft, under_review, needs_revision, accepted, accepted_with_reservations, rejected, superseded]
        changesFromPrevious:
          type: array
          items: { type: string }
          nullable: true
        createdBecauseOfCritiqueIds:
          type: array
          items: { type: string }
          nullable: true
        createdAt: { type: string, format: date-time }
    Claim:
      type: object
      required: [id, projectId, ideaVersionId, text, type, requiresEvidence, criticality, status, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        ideaVersionId: { type: string }
        text: { type: string }
        type:
          type: string
          enum: [technical, product, market, business, legal, ux, research, risk, assumption]
        requiresEvidence: { type: boolean }
        criticality:
          type: string
          enum: [low, medium, high, blocking]
        status:
          type: string
          enum: [unverified, supported, partially_supported, contradicted, unsupported, needs_external_validation]
        confidence: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
...
    CreateIdeaVersionRequest:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        changesFromPrevious:
          type: array
          items: { type: string }
        createdBecauseOfCritiqueIds:
          type: array
          items: { type: string }
    ExtractClaimsRequest:
      type: object
      properties:
        modelId: { type: string, nullable: true }
        force: { type: boolean, default: false }
...
    IdeaVersionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/IdeaVersion'
    IdeaVersionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/IdeaVersion'
    IdeaVersionDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            ideaVersion:
              $ref: '#/components/schemas/IdeaVersion'
            claims:
              type: array
              items:
                $ref: '#/components/schemas/Claim'
            critiques:
              type: array
              items: { type: object, additionalProperties: true }
...
    ClaimListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Claim'
...
    UserPresence:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string, nullable: true }
        userName: { type: string }
        page: { type: string }
        lastSeenAt: { type: string, format: date-time }
    PresenceListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            presence:
              type: array
              items:
                $ref: '#/components/schemas/UserPresence'
...